+++ 2025/hash.h
@@ -15,15 +15,15 @@ typedef struct {
 
 typedef struct {
 	Entry *entries;
-	size_t capacity;
-	size_t size;
+	uint64_t capacity;
+	uint64_t size;
 } HashTable;
 
 
 
 uint64_t hash(uint64_t x);
 uint64_t hashKey(Key key);
-void initializeHashTable(HashTable *table, size_t capacity);
+void initializeHashTable(HashTable *table, uint64_t capacity);
 void freeHashTable(HashTable *table);
 void insertKey(HashTable *table, Key key, uint64_t value);
 bool lookUpKey(HashTable *table, Key key, uint64_t *out);
